requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
tensorflow>=2.13.0
keras>=2.13.0
mtcnn>=1.0.0
numpy>=1.24.0,<2.0.0
pillow>=9.0.0
